requirements.txt
alabaster==0.7.10
anaconda-client==1.6.9
anaconda-navigator==1.7.0
anaconda-project==0.8.2
argh==0.26.2
asn1crypto==0.24.0
astroid==1.6.1
astropy==3.0
attrs==17.4.0
Babel==2.5.3
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
bitarray==0.8.1
bkcharts==0.2
blaze==0.11.3
bleach==2.1.3
bokeh==0.12.14
boto==2.48.0
Bottleneck==1.2.1
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
click==6.7
cloudpickle==0.5.2
clyent==1.2.2
colorama==0.3.9
conda==4.4.10
conda-build==3.4.1
conda-verify==2.0.0
contextlib2==0.5.5
cryptography==2.1.4
cycler==0.10.0
Cython==0.27.3
cytoolz==0.9.0.1
dask==0.17.1
datashape==0.5.4
decorator==4.2.1
distributed==1.21.3
dlib==19.12.0
docopt==0.6.2
docutils==0.14
EasyProcess==0.2.3
entrypoints==0.2.3
et-xmlfile==1.0.1
fastcache==1.0.2
filelock==3.0.4
Flask==0.12.2
Flask-Cors==3.0.3
gevent==1.2.2
gitdb2==2.0.3
GitPython==2.1.8
glob2==0.6
gmpy2==2.0.8
greenlet==0.4.13
h5py==2.7.1
heapdict==1.0.0
html5lib==1.0.1
idna==2.6
imageio==2.2.0
imagesize==1.0.0
imutils==0.4.6
ipykernel==4.8.2
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.2
isort==4.3.4
itsdangerous==0.24
jdcal==1.3
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.2
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyterlab==0.31.12
jupyterlab-launcher==0.10.5
Kivy==1.10.1.dev0
Kivy-Garden==0.1.4
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
llvmlite==0.22.0
locket==0.2.0
lxml==4.1.1
MarkupSafe==1.0
matplotlib==2.2.0
mccabe==0.6.1
mistune==0.8.3
mpmath==1.0.0
msgpack==0.5.6
msgpack-python==0.5.6
multipledispatch==0.5.0
navigator-updater==0.1.0
nbconvert==5.3.1
nbformat==4.4.0
neovim==0.2.4
networkx==2.1
nltk==3.2.5
nose==1.3.7
notebook==5.4.0
numba==0.37.0
numexpr==2.6.4
numpy==1.14.1
numpydoc==0.7.0
odo==0.5.1
olefile==0.45.1
opencv-python==3.4.1.15
openpyxl==2.5.1
packaging==17.1
pandas==0.22.0
pandocfilters==1.4.2
parso==0.1.1
partd==0.3.8
path.py==11.0
pathlib2==2.3.0
pathtools==0.1.2
patsy==0.5.0
pep8==1.7.1
pexpect==4.4.0
pickleshare==0.7.4
Pillow==5.0.0
pkginfo==1.4.1
pluggy==0.6.0
ply==3.11
prompt-toolkit==1.0.15
psutil==5.4.3
ptyprocess==0.5.2
py==1.5.2
pycodestyle==2.3.1
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0.1
pyflakes==1.6.0
Pygments==2.2.0
pylint==1.8.2
pyodbc==4.0.22
PyOpenGL==3.1.0
pyOpenSSL==17.5.0
pyparsing==2.2.0
Pypubsub==4.0.0
PyQt5==5.9.2
pyscreenshot==0.4.2
PySocks==1.6.8
pytest==3.4.2
python-dateutil==2.7.0
python-xlib==0.23
pytz==2018.3
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==17.0.0
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.4.0
requests==2.18.4
rope==0.10.7
ruamel-yaml==0.15.35
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.0.0
seaborn==0.8.1
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
sip==4.19.8
six==1.11.0
smmap2==2.0.3
snowballstemmer==1.2.1
sortedcollections==0.5.3
sortedcontainers==1.5.9
Sphinx==1.7.1
sphinxcontrib-websupport==1.0.1
spyder==3.2.7
SQLAlchemy==1.2.5
statsmodels==0.8.0
sympy==1.1.1
tables==3.4.2
tblib==1.3.2
terminado==0.8.1
testpath==0.3.1
toolz==0.9.0
tornado==5.0
traitlets==4.3.2
typing==3.6.4
unicodecsv==0.14.1
urllib3==1.22
watchdog==0.8.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.1.4
wrapt==1.10.11
wxPython==4.0.1
xlrd==1.1.0
XlsxWriter==1.0.2
xlwt==1.3.0
xonsh==0.6.3
zict==0.1.3
